[core] Rename commandhandler to command_handler, remove async command logging
